test(scale-down): move EC2 integration tests
